refactor: 캠퍼스 운영 규칙 관련 책임 분리
